+++ src/main/java/org/sonar/plugins/pitest/metrics/PitestMetricsKeys.java
@@ -17,7 +17,7 @@
  * License along with this program; if not, write to the Free Software
  * Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02
  */
-package org.sonar.plugins.pitest;
+package org.sonar.plugins.pitest.metrics;
 
 /**
  * {@link PitestMetrics} keys.
