+++ src/main/java/org/sonar/plugins/pitest/ui/PitestDashboardWidget.java
@@ -17,29 +17,36 @@
  * License along with this program; if not, write to the Free Software
  * Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02
  */
-package org.sonar.plugins.pitest;
+package org.sonar.plugins.pitest.ui;
 
-import java.util.List;
-
-import org.sonar.api.batch.AbstractSumChildrenDecorator;
-import org.sonar.api.batch.DependedUpon;
-import org.sonar.api.measures.Metric;
+import org.sonar.api.web.AbstractRubyTemplate;
+import org.sonar.api.web.Description;
+import org.sonar.api.web.NavigationSection;
+import org.sonar.api.web.RubyRailsWidget;
+import org.sonar.api.web.UserRole;
+import org.sonar.api.web.WidgetCategory;
 
 /**
- * Sum children decorator for pitest quantitative metrics.
- * 
+ * Sonar user widget for pitest metrics.
+ *   
  * @author <a href="mailto:aquiporras@gmail.com">Jaime Porras L&oacute;pez</a>
  */
-public class PitestDecorator extends AbstractSumChildrenDecorator {
+@NavigationSection(NavigationSection.RESOURCE)
+@UserRole(UserRole.USER)
+@WidgetCategory("Pitest")
+@Description("Pitest mutation coverage report.")
+public class PitestDashboardWidget extends AbstractRubyTemplate implements RubyRailsWidget {
 
-	@DependedUpon
-	@Override
-	public List<Metric> generatesMetrics() {
-		return PitestMetrics.getQuantitativeMetrics();
+	public String getId() {
+		return "pitest";
+	}
+
+	public String getTitle() {
+		return "Pitest report";
 	}
 
 	@Override
-	protected boolean shouldSaveZeroIfNoChildMeasures() {
-		return false;
+	protected String getTemplatePath() {
+		return "/org/sonar/plugins/pitest/pitest_dashboard_widget.html.erb";
 	}
 }
